requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
celery==5.5.3
redis==7.1.0

